# api/src/api/app/models/compartment_aggregation.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from app.models.tag import Tag

class CompartmentAggregation(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    CompartmentAggregation - a model defined in OpenAPI

        name: The name of this CompartmentAggregation [Optional].
        description: The description of this CompartmentAggregation [Optional].
        tags: The tags of this CompartmentAggregation [Optional].
    """
    # id :Optional[str] = Field(alias="id", default=None)
    name: Optional[str] = Field(alias="name", default=None)
    description: Optional[str] = Field(alias="description", default=None)
    tags: Optional[List[str]] = Field(alias="tags", default=None)

CompartmentAggregation.update_forward_refs()