# api/src/api/app/models/new_model.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from app.models.compartment import Compartment


class NewModel(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    NewModel - a model defined in OpenAPI

        name: The name of this NewModel [Optional].
        description: The description of this NewModel [Optional].
        aggregations: The aggregations of this NewModel [Optional].
        compartments: The compartments of this NewModel.
        groups: The groups of this NewModel.
        parameter_definitions: The parameter_definitions of this NewModel.
    """

    name: Optional[str] = Field(alias="name", default=None)
    description: Optional[str] = Field(alias="description", default=None)
    aggregations: Optional[List[List]] = Field(alias="aggregations", default=None)
    compartments: List[Compartment] = Field(alias="compartments")
    groups: List[List] = Field(alias="groups")
    parameter_definitions: List[List] = Field(alias="parameterDefinitions")

NewModel.update_forward_refs()