# api/src/api/app/models/node.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class Node(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    Node - a model defined in OpenAPI

        name: The name of this Node [Optional].
        description: The description of this Node [Optional].
        id: The id of this Node.
        ags: The ags of this Node [Optional].
    """

    name: Optional[str] = Field(alias="name", default=None)
    description: Optional[str] = Field(alias="description", default=None)
    id: str = Field(alias="id")
    ags: Optional[str] = Field(alias="AGS", default=None)

Node.update_forward_refs()